First implementation of the peripheral trait
This is the most basic implementation I could imagine. A `Peripheral` is a trait that is
implemented on a struct and it only provides a method to retrieve the properties of the peripheral.

An example library called `hello-world.rs` is now included to demonstrate how a peripheral plugin
might use this trait.