Add a plugin to control a single GPIO pin
A new plugin crate is added called kpal-gpio-cdev that enables the KPAL daemon
to control the output of GPIO pin 4 on a Raspberry Pi.

To enable this plugin, the error handling of the kpal-plugin library was
significantly refactored to allow ergonomic error handling inside individual
plugins while still forcing the plugin FFI layer to follow a strict interface.