Add a KpalLibraryInit type to kpal-plugin
This fixes the problem in issue-15 in which the daemon was using a hard-coded
function signature for the function that initializes a library. This improvement
means that the the signature can now change without breaking the daemon code.

In addition, cargo clippy is now run against the correct target on the CI.