Added callbacks to the Attribute type in the kpal-plugin library
This addition improves the ergonomics of writing plugins because code that is
specific to communicating with hardware can be isolated inside these
callbacks.

In addition, there are now default implementations for the main PluginAPI
methods. This means that they no longer need to be defined inside plugin
libraries, which makes individual plugins much less verbose and allows authors
to focus on the plugin logic rather than integrating with KPAL.

Finally, docstrings were added to the gpio-cdev plugin.