Add preinit attributes
This is a new feature that allows users to set the value of attributes before a
plugin is initialized. It improves the ergonomics of writing plugins because it
allows the user to configure certain aspects of a plugin, such as device files,
without hardcoding values into a plugin.

- Add attributes to the Library model
- Rename methods inside init::libraries module
- Remove the peripheral field from the executor

I did this because I want to use the executor to interact with plugins where a
peripheral is not needed. For example, if I want to discover the attributes of a
plugin to use with a Library model, then it makes no sense to have a Peripheral
model instantiated.

- Initialize new library models with the corresponding plugin attributes
- Added callbacks for the plugin initialization phase
- Renamed constants module to error_codes in kpal-pluginn
- Add plugin lifecycle phase to attribute get and sets
- Renamed kpal_plugin_init to kpal_plugin_new
- Add a plugin_init function to the plugin API
- Add a PluginAPI call to determine whether attributes are pre-init
- Add method to executor to determine whether an attribute is pre-init
- Add pre-init field to attributes
- Accept data for peripheral attributes from user input
- An error is raised if non-preinit attributes are set during a post
- Pre-init post data may be optionally provided
- Changed Set callback to Update and removed its data members
- Save the plugin lifetime phase in the executor
- Added sync method to the executor to set pre-init attributes
- Use a MultiMap for a plugin's Attribute collection

I made this change to make it easier for plugin maintainers to refer to their
own plugins' attributes. Previously, the collection was a Vec, which meant that
it could only be indexed as an integer.

Now, an attribute may be referenced by either index or a descriptive string
name.

- Refactor plugins to use the new multimap
- Fix broken reference error in the executor sync method