Fix issue 20: infinite loop during initialization
This change fixes an infinite loop that could occur in the daemon when a plugin
is initialized and does not have any attributes. In addition, the implicit
requirement that users assign attribute IDs that start at 0 and increase by 1 is
now removed; users can set an attribute ID in a plugin to any unique, positive
integer.

To realize this, two new methods were added to the plugin API: `attribute_count`
and `attribute_ids`. These are used to retrieve the set of attribute IDs from a
plugin.