Issue-34 Make matching on all pairs of enum variants explicit
This is a tech-debt ticket that addresses a problem where matching on pairs of
enums would require remembering to add a new pair to the match without the
compiler being able to warn you of an error.

Now all pairs are made explicit so that the compiler will warn of a forgotten
match.

- Clean up the set_attribute_value PluginAPI function
- Make variants in eq function for Attribute explicit